+++ src/main/java/br/com/afroglow/backendAfroGlow/Repository/BoxRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import br.com.afroglow.backendAfroGlow.Models.Box;
 
 @Repository
-public interface BoxRepository extends JpaRepository<Box, Long> {
+public interface BoxRepository extends JpaRepository<Box, Integer> {
 }
