+++ src/main/java/br/com/afroglow/backendAfroGlow/Repository/DefinicaoDoCabeloRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DefinicaoDoCabeloRepository extends JpaRepository<DefinicaoDoCabelo, Long> {
+public interface DefinicaoDoCabeloRepository extends JpaRepository<DefinicaoDoCabelo, Integer> {
     // Se necessário, você pode adicionar métodos personalizados aqui
 }
