+++ src/main/java/br/com/afroglow/backendAfroGlow/Repository/PedidoBoxRepository.java
@@ -3,5 +3,5 @@ package br.com.afroglow.backendAfroGlow.Repository;
 import br.com.afroglow.backendAfroGlow.Models.PedidoBox;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PedidoBoxRepository extends JpaRepository<PedidoBox, Long> {
+public interface PedidoBoxRepository extends JpaRepository<PedidoBox, Integer> {
 }
